% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{access_api}
\alias{access_api}
\title{Get access to BOX API}
\usage{
access_api(id, secret)
}
\arguments{
\item{id}{the id for box api}

\item{secret}{the secret for box api}
}
\description{
Get access to BOX API
}
