% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.by.day.R
\name{aggregate_by_day}
\alias{aggregate_by_day}
\title{Aggregate the raw data into an equally-spaced data frame}
\usage{
aggregate_by_day(dat, variable, ID, timestamp, agg_fun)
}
\arguments{
\item{dat}{Data in a data.frame format}

\item{variable}{Name of variable in the data set you would like to re-structure}

\item{ID}{Name of ID variable in the data set}

\item{timestamp}{Name of time variable in the data set}

\item{agg_fun}{Function used for aggregation. Default is sum.}
}
\description{
Aggregate the raw data into an equally-spaced data frame
}
