% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute_missForest}
\alias{impute_missForest}
\title{Impute missing value with missForest}
\usage{
impute_missForest(dat)
}
\arguments{
\item{dat}{merged dataset with filter passive/active/baseline data}
}
\value{
A data frame after missing value imputation
}
\description{
Impute missing value with missForest
}
