% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_missingness.R
\name{impute_quality_check}
\alias{impute_quality_check}
\title{Make plots for imputation result}
\usage{
impute_quality_check(
  filter_df,
  active_df,
  baseline_df,
  var_name,
  priority_miss_perc,
  n_round,
  seed,
  imp_func
)
}
\arguments{
\item{filter_df}{filtered passive data}

\item{active_df}{active data}

\item{baseline_df}{baseline data}

\item{var_name}{passive measurement variable name}

\item{priority_miss_perc}{Percentage of observations per subject to be labeled as missing;1)value between 0 and 1: same percentage for each subject;2)NA: automatically according to missing percentage of each subject}

\item{n_round}{number of times to randomly sample observations to be labeled}

\item{seed}{fix seed so that simulation results are reproduceable}

\item{imp_func}{the function of imputation method, available options now: #1) missForest :my_missForest; #2) moving average of 7 days: seven_moving_everage}
}
\value{
a data frame with imputation quality measurement metrics
}
\description{
Make plots for imputation result
}
