% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_filtered_passive}
\alias{load_filtered_passive}
\title{Load filtered passive data}
\usage{
load_filtered_passive()
}
\value{
A data frame of filtered passive data
}
\description{
Load filtered passive data
}
