% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_imputation.R
\name{plot_imputation}
\alias{plot_imputation}
\title{Make plots for imputation result}
\usage{
plot_imputation(raw_df, filter_df, impute_df, var_name, subject_name)
}
\arguments{
\item{raw_df}{raw passive dataframe}

\item{filter_df}{filtered passive dataframe}

\item{impute_df}{imputed passive dataframe}

\item{var_name}{passive measurement variable name}

\item{subject_name}{name of interested subject}
}
\value{
a ggplot2 object
}
\description{
Make plots for imputation result
}
